Move query language functions out of helpers
